rearrange migration service
