+++ administrator/com_joomgallery/language/en-GB/com_joomgallery.ini
@@ -738,6 +738,7 @@ COM_JOOMGALLERY_SERVICE_MIGRATION_MAPPING_SRC_IMAGETYPE_NOT_EXIST="The source im
 COM_JOOMGALLERY_SERVICE_MIGRATION_MAPPING_IMAGETYPE_NOT_USED="The following destination imagetypes are not used in the mapping: '%s'. Please go back to step 1 and apply the mapping correctly."
 COM_JOOMGALLERY_SERVICE_MIGRATION_ERROR_RECORD_ID_MISSING="Starting a migration without source data ID not possible. Contact the developer of the migration script to solve this."
 COM_JOOMGALLERY_SERVICE_ERROR_MIGRATION_RESUME="Migration can not be resumed. There is a problem in fetching the required info."
+COM_JOOMGALLERY_SERVICE_MIGRATION_FAILED_FETCH_DATA="Data from source could not be loaded. See log file for more destails."
 COM_JOOMGALLERY_SERVICE_MIGRATION_FAILED_CONVERT_DATA="Data from source could not be converted to new data structure. See log file for more destails."
 COM_JOOMGALLERY_SERVICE_MIGRATION_FAILED_INSERT_RECORD="Data could not be inserted into destination database. See log file for more destails."
 COM_JOOMGALLERY_SERVICE_MIGRATION_FAILED_CREATE_IMGTYPE="Error in creating imagetypes. See log file for more destails."
